Making ISIS working. Still requires a lot of namespace and xml cleaning.

+++ src/services/isis/centralisi.cpp
@@ -41,6 +41,13 @@ static void register_thread(void *arg)
     srv->reg->registration_forever();
 }
 
+static void MakeISIFault(Arc::XMLNode fault)
+{
+    fault.NewChild("isis:Name")="undefined";
+    fault.NewChild("isis:Description");
+    fault.NewChild("isis:Type")="1";
+}
+
 void CentralISIService::InformationCollector(void)
 {
     for (;;) {
@@ -48,6 +55,7 @@ void CentralISIService::InformationCollector(void)
         // Create infodoc
         Arc::NS ns;
         Arc::XMLNode node(ns, "InfoDoc");
+        // TODO: make rest Glue2
         Arc::XMLNode service_node = node.NewChild("Service");
         service_node.NewChild("ID") = service_id;
         service_node.NewChild("Name") = "CentralISIService";
@@ -56,7 +64,7 @@ void CentralISIService::InformationCollector(void)
         ep.NewChild("ID") = "0";
         ep.NewChild("URL") = "http://localhost:60000/isis";
         // Write data to info cache
-        if(icache) icache->Cache().Set("/", node);
+        if(icache) icache->Cache().Set("/local", node);
         sleep(60); // XXX: make it configurable
     }
 }
@@ -64,6 +72,7 @@ void CentralISIService::InformationCollector(void)
 CentralISIService::CentralISIService(Arc::Config *cfg):Service(cfg),logger(Arc::Logger::rootLogger, "CentralISI"),reg(NULL)
 {
     // Define supported namespaces
+    // TODO: Remove those addedd automatically and take others from headers
     ns["isis"]="http://www.nordugrid.org/schemas/isis/2007/06";
     ns["wsa"]="http://www.w3.org/2005/08/addressing";
     ns["wsrf-bf"]="http://docs.oasis-open.org/wsrf/bf-2";
@@ -72,6 +81,7 @@ CentralISIService::CentralISIService(Arc::Config *cfg):Service(cfg),logger(Arc::
     ns["wsrf-rw"]="http://docs.oasis-open.org/wsrf/rw-2";
     
     logger.msg(Arc::DEBUG, "Central ISIS initialized");
+    /*
     std::string dsn = (std::string)((*cfg)["DSN"]);
     if(dsn.empty()) {
         logger.msg(Arc::ERROR, "Missing database location in configuration");
@@ -101,10 +111,13 @@ CentralISIService::CentralISIService(Arc::Config *cfg):Service(cfg),logger(Arc::
             }        
         }
     }
-    Arc::XMLNode regx = (*cfg)["Register"];
+    */
+    Arc::Config regx((*cfg)["Register"]);
     service_id = (std::string)(*cfg)["ID"];
     if(service_id.empty()) service_id="ISIS";
-    icache = new Arc::InfoCacheInterface(*cfg, service_id);
+    Arc::Config icfg((*cfg)["InformationSystem"]);
+    icache = new Arc::InfoCacheInterface(icfg, service_id);
+    mcache = new Arc::InfoCache(icfg, service_id+".meta");
     CreateThreadFunction(&infocollector_thread, this);
     if(!regx) {
         logger.msg(Arc::WARNING, "Missing registration section, won't register anywhere");
@@ -112,7 +125,7 @@ CentralISIService::CentralISIService(Arc::Config *cfg):Service(cfg),logger(Arc::
         std::string s_reg_period = regx["Period"];
         long int reg_period = strtol(s_reg_period.c_str(), NULL, 10);
         if(reg_period == 0) reg_period=600;
-        reg = new Arc::InfoRegister(service_id, reg_period, (*cfg));
+        reg = new Arc::InfoRegister(service_id, reg_period, regx);
         reg->AddUrl("http://localhost:60000/isis");
         CreateThreadFunction(&register_thread, this);
     };
@@ -127,10 +140,12 @@ CentralISIService::~CentralISIService(void)
 Arc::MCC_Status CentralISIService::process(Arc::Message &inmsg, Arc::Message &outmsg)
 {
     logger.msg(Arc::DEBUG, "process called");
+    /*
     if (db_path == "") {
         logger.msg(Arc::ERROR, "invalid database path");
         return Arc::MCC_Status();
     }
+    */
     // Both input and output are supposed to be SOAP
     // Extracting payload
     Arc::PayloadSOAP* inpayload = NULL;
@@ -147,6 +162,9 @@ Arc::MCC_Status CentralISIService::process(Arc::Message &inmsg, Arc::Message &ou
       logger.msg(Arc::ERROR, "input does not define operation");
       return make_soap_fault(outmsg);
     }; 
+    std::string x_str;
+    op.GetXML(x_str);
+    logger.msg(Arc::VERBOSE, "SOAP request: %s",x_str);
     logger.msg(Arc::DEBUG, "process: operation: %s", op.Name());
     Arc::PayloadSOAP* outpayload = new Arc::PayloadSOAP(ns);
     Arc::PayloadSOAP& res = *outpayload;
@@ -190,6 +208,8 @@ Arc::MCC_Status CentralISIService::process(Arc::Message &inmsg, Arc::Message &ou
         return make_soap_fault(outmsg);
     };
     // Set output
+    outpayload->GetXML(x_str);
+    logger.msg(Arc::VERBOSE, "SOAP response: %s",x_str);
     outmsg.Payload(outpayload);
     return Arc::MCC_Status(ret);
 }
@@ -206,19 +226,56 @@ Arc::MCC_Status CentralISIService::make_soap_fault(Arc::Message& outmsg)
     return Arc::MCC_Status(Arc::STATUS_OK);
 }
 
-Arc::MCC_Status CentralISIService::Register(Arc::XMLNode &in, Arc::XMLNode &/*out*/)
+
+
+static void url_encode(std::string& s) {
+  for(std::string::size_type n=0;n<s.size();++n) {
+    if(isalnum(s[n])) continue;
+    char buf[4];
+    snprintf(buf,4,"%%%02x",(unsigned int)(s[n]));
+    buf[3]=0;
+    s.replace(n,1,buf,3); n+=2;
+  };
+}
+
+Arc::MCC_Status CentralISIService::Register(Arc::XMLNode &in, Arc::XMLNode &out)
 {
     int i;
     Arc::XMLNode entry;
-    // TODO: protection against malicious id like '../../../../../etc/passwd'
+    bool failed = false;
     for (i = 0; (bool)(entry = in["RegEntry"][i]); i++) {
-        std::string id = (std::string)entry["ID"];
+        // Meaning of element ID is undefined. Service is supposed to be 
+        // identified by EPR.
+        // std::string id = (std::string)entry["ID"];
+        std::string id = Arc::WSAEndpointReference(entry["SrcAdv"]["EPR"]).Address();
         if(id.empty()) {
-            logger.msg(Arc::ERROR, "Missing identifier in registration request");
+            // logger.msg(Arc::ERROR, "Missing identifier in registration request");
+            logger.msg(Arc::ERROR, "Missing address in registering service EPR");
+            failed=true; continue;
         }
-
-        if(icache) icache->Cache().Set("/"+id,entry);
-
+        // Converting id into safe string
+        url_encode(id);
+        Arc::XMLNode adv = entry["SrcAdv"];
+        if(!adv) {
+            logger.msg(Arc::ERROR, "Missing service advertisement in registering service");
+            failed=true; continue;
+        }
+        Arc::XMLNode madv = entry["MetaSrcAdv"];
+        if(!madv) {
+            logger.msg(Arc::ERROR, "Missing service meta advertisement in registering service");
+            failed=true; continue;
+        }
+        if(!icache) {
+            failed=true; continue;
+        }
+        if(!mcache) {
+            failed=true; continue;
+        }
+        Arc::XMLNode infodoc(ns,"InfoDoc");
+        infodoc.NewChild(adv);
+        failed|=!(icache->Cache().Set("/"+id,infodoc));
+        failed|=!(mcache->Set("/"+id,madv));
+        /*
         std::string data;
         entry.GetXML(data);
         std::string entry_file = Glib::build_filename(db_path, id);
@@ -229,7 +286,12 @@ Arc::MCC_Status CentralISIService::Register(Arc::XMLNode &in, Arc::XMLNode &/*ou
         out << "<?xml version=\"1.0\" encoding=\"utf-8\"?>" << std::endl;
         out << data;
         out.close();
+        */
     }
+    if(failed) {
+      logger.msg(Arc::ERROR, "Register: reporting fault");
+      MakeISIFault(out.NewChild("isis:Fault"));
+    };
     return Arc::MCC_Status(Arc::STATUS_OK);
 }
 
@@ -237,10 +299,19 @@ Arc::MCC_Status CentralISIService::RemoveRegistrations(Arc::XMLNode &in, Arc::XM
 {
     int i;
     Arc::XMLNode entry;
-    for (i = 0; (entry = in["ID"][i]) == true; i++) {
-        std::string id = (std::string)entry;
+    for (i = 0; (entry = in["RegEntryID"][i]) == true; i++) {
+        std::string id = Arc::WSAEndpointReference(entry).Address();
+        if(id.empty()) {
+            logger.msg(Arc::ERROR, "Missing address in unregistering service EPR");
+            continue;
+        }
+        url_encode(id);
+        if(icache) icache->Cache().Unset("/"+id);
+        if(mcache) mcache->Unset("/"+id);
+        /*
         std::string path = Glib::build_filename(db_path, id);
         unlink(path.c_str());
+        */
     }
     return Arc::MCC_Status(Arc::STATUS_OK);
 }
@@ -249,8 +320,32 @@ Arc::MCC_Status CentralISIService::GetRegistrationStatuses(Arc::XMLNode &in, Arc
 {
     int i;
     Arc::XMLNode entry;
-    for (i = 0; (entry = in["ID"][i]) == true; i++) {
-        std::string id = (std::string)entry;
+    for (i = 0; (entry = in["RegEntryID"][i]) == true; i++) {
+        std::string id = Arc::WSAEndpointReference(entry).Address();
+        Arc::XMLNode re = out.NewChild("isis:RegEntryStatus");
+        if(id.empty()) {
+            logger.msg(Arc::ERROR, "Missing address in service EPR");
+            continue;
+        } else {
+          url_encode(id);
+          if(mcache) {
+            Arc::XMLNodeContainer result;
+            if(mcache->Get("/"+id,result)) {
+              if(result.Size() >= 1) {
+                Arc::XMLNode status = result[0]["Status"];
+                if(status) {
+                  re.NewChild(entry);     // isis:RegEntryID
+                  re.NewChild("isis:ID"); // isis::ID
+                  re.NewChild(status);    // isis::Status
+                };
+              };
+            };
+          };
+        };
+        if(re.Size() == 0) {
+          MakeISIFault(re.NewChild("isis:FaultType"));
+        };
+        /*
         std::string path = Glib::build_filename(db_path, id);
         if (Glib::file_test(path, Glib::FILE_TEST_EXISTS) == true) {
             std::string data = Glib::file_get_contents(path);
@@ -262,6 +357,7 @@ Arc::MCC_Status CentralISIService::GetRegistrationStatuses(Arc::XMLNode &in, Arc
                 out["RegEntry"].NewChild(status);
             }
         }
+        */
     }
     std::string str;
     out.GetXML(str);
